crimson/os: dissect sharding managers into sharding_helpers.h
